Mobile: Add plugin panel integration test
