+++ packages/app-mobile/utils/testing/createMockReduxStore.ts
@@ -1,15 +1,16 @@
-import reducer from '@joplin/lib/reducer';
 import { createStore } from 'redux';
 import appDefaultState from '../appDefaultState';
 import Setting from '@joplin/lib/models/Setting';
 import { AppState } from '../types';
+import appReducer from '../appReducer';
 
 const testReducer = (state: AppState|undefined, action: unknown): AppState => {
 	state ??= {
 		...appDefaultState,
 		settings: Setting.toPlainObject(),
 	};
-	return { ...state, ...reducer(state, action) };
+
+	return { ...state, ...appReducer(state, action) };
 };
 
 const createMockReduxStore = () => {
